Revert "拆分不同楼层的线路，优化大厅到LG1-办公层的线路"
This reverts commit c9e62e97e5e8c83bb0538e8a2e726cc82a51c2b6.
