--- src/heist_path/LG1GotoLG2Path.py
@@ -1,336 +1,0 @@
-import time
-from src.heist_path.HeistPath import HeistPath
-
-# 路径参考自 B站 UP 主：@早柚大魔王
-class LG1GotoLG2Path(HeistPath):
-    def run_path(self):
-        self.switch_to_runner(check_switched=True)
-        self.lg1_wp1()
-        self.lg1_wp2()
-        self.lg1_wp3()
-        self.lg1_wp4()
-        idx = self.avoider_strategy_index()
-        if idx == -1:
-            self.lg1_wp5_avoid_combat_01()
-        elif idx == 0:
-            self.lg1_wp5_avoid_combat_02()
-        elif idx == 1:
-            self.lg1_wp5_avoid_combat_03()
-    
-    def lg1_wp1(self):
-        self.sleep(0.75)
-        self.send_key_down("w")
-        self.sleep(9.06)
-        self.send_key_up("w")
-        self.sleep(0.51)
-        self.send_key_down("d")
-        self.sleep(1.71)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(1.01)
-        self.wait_and_interact(direction="s")
-        self.sleep(0.11)
-        self.send_key_down("f")  # start pick
-        self.sleep(0.78)
-        self.send_key_down("s")
-        self.sleep(1.25)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(3.03)
-        self.send_key_up("d")
-        self.sleep(0.22)
-        self.send_key_down("a")
-        self.sleep(3.90)
-        self.send_key_up("a")
-        self.sleep(0.31)
-        self.send_key_down("d")
-        self.sleep(0.40)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(2.01)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(5.60)
-        self.send_key_up("d")
-        self.sleep(0.06)
-        self.send_key_down("w")
-        self.sleep(2.02)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(3.21)
-        self.send_key_up("d")
-        self.sleep(0.12)
-
-    def lg1_wp2(self):
-        self.send_key_down("d")
-        self.sleep(1.80)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(0.71)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.71)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(2.28)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.26)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_up("f")  # end pick
-        self.sleep(2.93)
-        self.send_key_down("w")  # 过镭射1
-        self.sleep(2.01)
-        self.send_key_up("w")
-        self.sleep(0.44)
-        self.start_interaction_watch()
-        self.send_key_down("w")  # 过镭射2
-        self.sleep(8.51)
-        self.send_key_up("w")
-        self.stop_interaction_watch()
-        self.sleep(0.33)
-
-    def lg1_wp3(self):
-        self.send_key_down("f")  # start pick
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(2.13)
-        self.send_key_up("a")
-        self.sleep(0.52)
-        self.send_key_down("s")
-        self.sleep(1.32)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(0.20)
-        self.send_key_up("w")
-        self.sleep(0.31)
-        self.send_key_down("a")
-        self.sleep(1.50)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.20)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.start_interaction_watch()
-        self.send_key_down("w")
-        self.sleep(3.19)
-        self.send_key_down("a")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(5.16)
-        self.send_key_up("w")
-        self.stop_interaction_watch()
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.15)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_up("f")  # end pick
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(2.51)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.40)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(5.31)
-        self.send_key_up("w")
-        self.sleep(0.12)
-
-    def lg1_wp4(self):
-        self.send_key_down("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(3.31)
-        self.send_key_up("s")
-        self.sleep(0.12)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.50)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.start_interaction_watch()
-        self.send_key_down("f")  # start pick
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.11)
-        self.send_key_up("a")
-        self.sleep(1.22)
-        self.stop_interaction_watch()
-        self.send_key_down("w")
-        self.sleep(6.58)
-        self.send_key_down("d")
-        self.sleep(2.62)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_down("a")
-        self.sleep(0.31)
-        self.send_key_up("a")
-        self.sleep(0.32)
-        self.send_key_down("w")
-        self.sleep(0.21)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.31)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(1.30)
-        self.start_interaction_watch()
-        self.send_key_down("d")
-        self.sleep(2.10)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.sleep(0.65)
-        self.send_key_down("w")
-        self.sleep(0.22)
-        self.send_key_down("d")
-        self.sleep(0.61)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.56)
-        self.send_key_down("space")
-        self.sleep(0.06)
-        self.send_key_up("space")
-        self.sleep(0.14)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_up("w")
-        self.sleep(0.34)
-        self.send_key_down("d")
-        self.sleep(1.41)
-        self.stop_interaction_watch()
-        self.send_key_down("w")
-        self.sleep(0.81)
-        self.send_key_up("w")
-        self.sleep(0.11)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("s")
-        self.sleep(0.47)
-        self.send_key_up("s")
-        self.sleep(0.11)
-        self.send_key_down("d")
-        self.sleep(0.60)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(3.38)
-        self.send_key_up("w")
-        self.sleep(0.34)
-        self.send_key_down("d")
-        self.sleep(0.61)
-        self.send_key_up("d")
-        self.sleep(0.11)
-        self.loot_safes_while_walking(direction="s", time_out=2.37)
-        self.sleep(0.11)
-        self.send_key_up("f")  # end pick
-        self.sleep(0.10)
-        self.send_key_down("d")
-        self.sleep(1.33)
-        self.send_key_up("d")
-        self.sleep(0.12)
-        self.send_key_down("w")
-        self.sleep(9.40)
-        self.send_key_up("w")
-        self.sleep(0.31)
-
-    def lg1_wp5_avoid_combat_01(self):
-        self.send_key_down("w")
-        self.sleep(2.02)
-        self.send_key_up("w")
-        self.sleep(0.10)
-        self.send_key_down("s")
-        self.sleep(0.11)
-        self.send_key_down("lshift")
-        self.sleep(0.06)
-        self.send_key_up("lshift")
-        self.sleep(0.81)
-        self.send_key_up("s")
-        self.sleep(2.01)
-        self.send_key_down("w")
-        self.sleep(0.11)
-
-        deadline = time.time() + 4.5
-        while time.time() < deadline:
-            self.send_key("lshift")
-            self.sleep(0.51)
-
-        self.wait_and_interact(direction="w", is_lock=True)
-        self.sleep(0.11)
-        self.switch_to_runner(check_switched=True)
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.00)
-        self.wait_and_interact(direction="w")
-
-    def lg1_wp5_avoid_combat_02(self):
-        self.send_key_down("s")
-        self.sleep(1.50)
-        self.send_key_up("s")
-        self.sleep(0.11)
-
-        self.switch_to_avoider(check_switched=True)
-        self.sleep(0.5)
-        self.perform_avoidance_action()
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(6.0)
-        self.send_key_up("w")
-        self.switch_to_runner(check_switched=True)
-        self.sleep(0.5)
-        self.wait_and_interact(is_lock=True)
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.00)
-        self.wait_and_interact(direction="w")
-
-    def lg1_wp5_avoid_combat_03(self):
-        self.switch_to_avoider(check_switched=True)
-        self.sleep(0.5)
-        self.perform_avoidance_action()
-        self.sleep(3.2)
-        self.send_key_down("w")
-        self.sleep(0.11)
-
-        deadline = time.time() + 4.5
-        while time.time() < deadline:
-            self.send_key("lshift")
-            self.sleep(0.51)
-
-        self.send_key_up("w")
-        self.sleep(0.2)
-        self.perform_avoidance_action()
-        self.sleep(3.2)
-        self.switch_to_runner(check_switched=True)
-        self.sleep(0.5)
-        self.wait_and_interact(is_lock=True)
-        self.sleep(0.11)
-        self.send_key_down("w")
-        self.sleep(1.00)
-        self.wait_and_interact(direction="w")
